requirements.txt
# Top-level Python requirements for prodmgr
# (C) Copyright 2021-2023 Hewlett Packard Enterprise Development LP.

pyyaml
